refactor: splitting the solution into multiple files/functions (from solution.js into read-write-files.js and crack-caesar-cipher.js)
