+++ challenges/intermediate/crack-caesar-cipher/solutions/aleattene/encode_decode.js
@@ -23,6 +23,8 @@ function decode(encryptedMessage, shift) {
     return(encode(encryptedMessage, -shift));
 }
 
+
+// Export modules
 module.exports = {
     encode,
     decode
